Fix unknown pdeathsig field on darwin os
